rename harness/assert.{c,h} to harness/sim_assert.{c,h}
- the previous 'assert.h' shadowed the system <assert.h> via the
  -I harness include path and silently broke any translation unit
  that needed the standard assert(). Renaming removes the trap
  entirely and matches the sim_expect_* naming already used inside
  the file.
